feat(socker): create handler for sockets
