+++ go/api/handlers/chatMessage.go
@@ -8,11 +8,8 @@ import (
 	"challenge-flutter-go/logger"
 	"challenge-flutter-go/models"
 	"challenge-flutter-go/repository"
-	"fmt"
 	"net/http"
 	"time"
-
-	"github.com/gorilla/websocket"
 
 	"github.com/gin-gonic/gin"
 )
@@ -22,14 +19,14 @@ type ChatMessageHandler struct {
 	TripRepository repository.TripRepository
 }
 
-var upgrader = websocket.Upgrader{
-	CheckOrigin: func(r *http.Request) bool {
-		return true
-	},
-}
+// var upgrader = websocket.Upgrader{
+// 	CheckOrigin: func(r *http.Request) bool {
+// 		return true
+// 	},
+// }
 
-var clients = make(map[*websocket.Conn]bool)
-var broadcast = make(chan responses.ChatMessageResponse)
+// var clients = make(map[*websocket.Conn]bool)
+// var broadcast = make(chan responses.ChatMessageResponse)
 
 // @Summary		Create a new chat message on trip
 // @Description	Create a new chat message & associate it with the trip
@@ -139,22 +136,5 @@ func (handler *ChatMessageHandler) HandleConnections(c *gin.Context) {
 			break
 		}
 		broadcast <- msg
-	}
-}
-
-func (handler *ChatMessageHandler) HandleMessages() {
-	fmt.Println("Handle messages")
-	for {
-		msg := <-broadcast
-		fmt.Println("Received message :", msg)
-		fmt.Println("Broadcasting message")
-		for client := range clients {
-			fmt.Println("Sending message to client")
-			err := client.WriteJSON(msg)
-			if err != nil {
-				client.Close()
-				delete(clients, client)
-			}
-		}
 	}
 }
